Add files via upload
This is a CAD model made using the software Solidworks, to attach the Power bank, Raspberry pi and web-camera onto the walker. This design will be 3D printed using PLA (Polylactic Acid filament), a renewable, recyclable and natural material. The Power bank, raspberry pi, metal rods and web-camera are not a part of the model, they were added to the assembly to make sure the design is correct and to envision the entire product. The CAD model for web-camera and raspberry pi were found online, other than that everything was made from scratch.  This design will be attached to the walker, and was made keeping in mind cost and material efficiency.